# docs/swagger.yaml
basePath: /v1
definitions:
...
  auth.initialiseRegisterRequest:
    properties:
      account_type:
        example: 1
        type: integer
      email:
        example: user@email.com
        type: string
      phone_number:
        example: "0712345678"
        type: string
    required:
    - account_type
    type: object
  auth.initialiseRegisterResponse:
    properties:
      message:
        type: string
    type: object
...
  auth.registerBusinessRequest:
    properties:
      business_name:
        example: John XYZ Plc
        type: string
      country_code:
        example: "234"
        type: string
      email:
        example: user@email.com
        type: string
      industry_type:
        example: 1
        type: integer
      password:
        example: password
        type: string
      phone_number:
        example: "0712345678"
        type: string
    required:
    - business_name
    - industry_type
    - password
    type: object
  auth.registerBusinessResponse:
    properties:
      user:
        $ref: '#/definitions/dtos.Business'
    type: object
  auth.registerResponse:
    properties:
      user:
        $ref: '#/definitions/dtos.User'
    type: object
  auth.validateOTPRequest:
    properties:
      email:
        example: user@email.com
        type: string
      otp:
        example: "123456"
        type: string
      phone_number:
        example: "0712345678"
        type: string
    required:
    - otp
    type: object
  auth.validateOTPResponse:
    properties:
      token:
        type: string
    type: object
...
  dtos.Business:
    properties:
      business_name:
        type: string
      country_code:
        type: string
      email:
        type: string
      industry_type:
        type: string
      phone_number:
        type: string
      user_id:
        type: string
    type: object
...
  dtos.User:
    properties:
      country_code:
        type: string
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      interests:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      relationship_status:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
...
  handlers.resendOTPRequest:
    properties:
      email:
        example: user@email.com
        type: string
      mode:
        example: register
        type: string
      phone_number:
        example: "0712345678"
        type: string
    required:
    - mode
    type: object
  handlers.resendOTPResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8083
info:
  contact: {}
  description: Celebut Backend REST endpoints
  title: Celebut API
  version: "1.0"
paths:
...
  /otp/resend:
    post:
      consumes:
      - application/json
      description: Resend OTP code
      operationId: resend-otp
      parameters:
      - description: Resend OTP
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.resendOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.resendOTPResponse'
      summary: Resend OTP
      tags:
      - OTP
...
  /register/business:
    post:
      consumes:
      - application/json
      description: Complete business registration
      operationId: register-business
      parameters:
      - description: Complete a new business registration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.registerBusinessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.registerBusinessResponse'
      security:
      - Bearer: []
      summary: Complete business registration
      tags:
      - Authentication
  /register/initialise:
    post:
      consumes:
      - application/json
      description: Begin registration process
      operationId: initialise-register
      parameters:
      - description: begin a new registration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.initialiseRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.initialiseRegisterResponse'
      security:
      - Auth-Token: []
      summary: Initialise Registration
      tags:
      - Authentication
  /register/user:
    post:
      consumes:
      - multipart/form-data
      description: Complete a user's registration
      operationId: register-user
      parameters:
      - description: profile image file
        in: formData
        name: profile_image
        type: file
      - example: "234"
        in: formData
        name: country_code
        type: string
      - example: "2020-10-12"
        in: formData
        name: dob
        required: true
        type: string
      - example: user@email.com
        in: formData
        name: email
        type: string
      - example: John
        in: formData
        name: first_name
        required: true
        type: string
      - example: male
        in: formData
        name: gender
        type: string
      - example: fashion,entertainment
        in: formData
        name: interests
        type: string
      - example: Doe
        in: formData
        name: last_name
        required: true
        type: string
      - example: password
        in: formData
        name: password
        required: true
        type: string
      - example: "0712345678"
        in: formData
        name: phone_number
        type: string
      - example: single
        in: formData
        name: relationship_status
        type: string
      - example: johndoe
        in: formData
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.registerResponse'
      security:
      - Bearer: []
      summary: Complete User Registration
      tags:
      - Authentication
  /register/validate:
    post:
      consumes:
      - application/json
      description: Validate registration OTP
      operationId: validate-register-otp
      parameters:
      - description: otp validation body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.validateOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.validateOTPResponse'
      security:
      - Auth-Token: []
      summary: Validate Registration OTP
      tags:
      - Authentication
...
securityDefinitions:
  Auth-Token:
    description: Type token for unauthenticated requests
    in: header
    name: x-auth-token
    type: apiKey
  Bearer:
    description: '`Bearer token` for authenticated requests'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"